api/requirements.txt
blinker
click
Flask
Flask-SQLAlchemy  # Adicione isso
Flask-CORS  # Adicione isso
greenlet
itsdangerous
Jinja2
MarkupSafe
psycopg2
python-dotenv
SQLAlchemy
typing_extensions
Werkzeug